% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomDistMatrix.R
\name{generateRandomDistMatrix}
\alias{generateRandomDistMatrix}
\title{Generate Random Distance Matrix}
\usage{
generateRandomDistMatrix(numCities, maxDistance = 100)
}
\arguments{
\item{numCities}{The number of cities.}

\item{maxDistance}{The maximum distance between cities. Default is 100.}
}
\value{
A symmetric matrix of distances between cities.
}
\description{
This function generates a random symmetric distance matrix for the Traveling Salesperson Problem.
}
\examples{
set.seed(123)
distMatrix <- generateRandomDistMatrix(5)
print(distMatrix)
}
