% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticAlgorithm.R
\name{geneticAlgorithm}
\alias{geneticAlgorithm}
\title{Genetic Algorithm for TSP}
\usage{
geneticAlgorithm(distMatrix, popSize = 50, maxGenerations = 100)
}
\arguments{
\item{distMatrix}{A distance matrix representing the distances between cities.}

\item{popSize}{The size of the population for the genetic algorithm. Default is 50.}

\item{maxGenerations}{The maximum number of generations for the genetic algorithm. Default is 100.}
}
\value{
A list containing the best tour and the total distance of the tour.
}
\description{
This function solves the Traveling Salesperson Problem using a genetic algorithm.
}
\examples{
distMatrix <- matrix(c(
  0, 29, 20, 21,
  29, 0, 15, 17,
  20, 15, 0, 28,
  21, 17, 28, 0
), nrow = 4, byrow = TRUE)
result <- geneticAlgorithm(distMatrix)
print(result)
}
