% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbor.R
\name{nearestNeighbor}
\alias{nearestNeighbor}
\title{Nearest Neighbor TSP Algorithm}
\usage{
nearestNeighbor(distMatrix)
}
\arguments{
\item{distMatrix}{A distance matrix representing the distances between cities.}
}
\value{
A list containing the tour and the total distance of the tour.
}
\description{
This function solves the Traveling Salesperson Problem using the nearest neighbor heuristic.
}
\examples{
distMatrix <- matrix(c(
  0, 29, 20, 21,
  29, 0, 15, 17,
  20, 15, 0, 28,
  21, 17, 28, 0
), nrow = 4, byrow = TRUE)
result <- nearestNeighbor(distMatrix)
print(result)
}
