% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveTSP.R
\name{solveTSP}
\alias{solveTSP}
\title{Solve TSP Using Multiple Methods}
\usage{
solveTSP(distMatrix)
}
\arguments{
\item{distMatrix}{A distance matrix representing the distances between cities.}
}
\value{
A list containing the best tour and the total distance of the best tour found by the algorithms.
}
\description{
This function solves the Traveling Salesperson Problem using multiple heuristic methods and returns the best solution found.
}
\examples{
distMatrix <- matrix(c(
  0, 29, 20, 21,
  29, 0, 15, 17,
  20, 15, 0, 28,
  21, 17, 28, 0
), nrow = 4, byrow = TRUE)
result <- solveTSP(distMatrix)
print(result)
}
